мониторинг ГК + removed unused usings

+++ Projects/FireMonitor/Infrastructure/Events/ShowXZoneEvent.cs
@@ -2,7 +2,7 @@
 
 namespace Infrastructure.Events
 {
-    public class ShowXZoneEvent : CompositePresentationEvent<ulong?>
+    public class ShowXZoneEvent : CompositePresentationEvent<short?>
     {
     }
 }
